OCPBUGS-15253: Include ingress's namespace in ingress without class metric
OpenShift's ingress operator creates a prometheus alert based on the
openshift_ingress_to_route_controller_ingress_without_class_name metric,
and to make it clearer which ingress is missing an ingress class name,
the alert message needs to include ingress's namespace.  This adds the
namespace as a tag on the metric, which will be used in the alert
message.