ingress: Reset metrics when ingress is deleted
When an ingress object is deleted, reset the gauge for that ingress object to 0
in order to prevent stale metrics from triggering alerts.

This commit fixes OCPBUGS-13181.

https://issues.redhat.com/browse/OCPBUGS-13181

* pkg/route/ingress/ingress.go ((Controller).sync): Call ResetIngressMetrics if
the ingress is not found.
* pkg/route/ingress/metrics.go ((Controller).ResetIngressMetrics): New function.
Reset the openshift_ingress_to_route_controller_ingress_without_class_name gauge
for the specified ingress to 0.
* pkg/route/ingress/metrics_test.go (TestMetrics): Unregister the collector when
the test is done so that subsequent tests can register it.
(Test_ResetIngressMetrics): Verify that ResetIngressMetrics has the intended
effect.